# Pushdown  single column predicates from ON join clauses