… benchmarks/expected-plans/q19.txt
… datafusion/core/tests/sql/predicates.rs