@alamb 2022-10-12
I would be willing to help write some more tests / comments for this PR if others think it is a good idea.