@AssHero 2022-10-13
> @AssHero do you have some numbers, e.g. how much does it help q19?
> 
> @alamb one thing we might consider is to apply the optimization whenever there is an expensive parent node (joins/aggregate) that benefits from more filtering? Anyway, in most cases, filtering should be cheap compared to other operations as long as the expression is not very expensive (for example, no expensive UDF).

Before this commit 'Convert more cross joins to inner joins'(https://github.com/apache/arrow-datafusion/pull/3482)， this optimization helps much in q19, and I intruduce this optimization to solve q19's performance problem. 
Now q19 is transformed to inner join, and this helps little in q19, but I think this will help other queries with OR clauses.