@AssHero 2022-10-13
> I think this optimization might actually slow some plans down (if the filters that are pushed down don't actually filter many rows, they will just consume CPU). I don't have a great suggestion about this

I have made some tests on q19 with this optimization on current version, it does not slow down the queries even through it helps little(may consume more CPU). 
I think more better way is combined with statistic data(CBO).  With the statistic data, we can evaluate the rows filtered by this predicate, then make the decision.