@AssHero 2022-10-13
add some comments about extract_or_clause.