@alamb 2022-10-15
The newly added TPCH plan benchmarks needed to be updated as well