@Dandandan 2022-10-15
@alamb running the benchmarks now for q7 and q19 posting them here when done