@Dandandan 2022-10-15
Looks like performance regressed a bit :/ we probably have a better look before merging it @alamb 

Master:

```
Query 7 iteration 0 took 6961.0 ms and returned 4 rows
Query 7 iteration 1 took 6460.2 ms and returned 4 rows
Query 7 iteration 2 took 6454.0 ms and returned 4 rows
Query 7 avg time: 6625.06 ms

Query 19 iteration 0 took 1305.5 ms and returned 1 rows
Query 19 iteration 1 took 1046.6 ms and returned 1 rows
Query 19 iteration 2 took 1053.0 ms and returned 1 rows
Query 19 avg time: 1135.04 ms
```

New:

```
Query 7 iteration 0 took 10860.9 ms and returned 4 rows
Query 7 iteration 1 took 9867.5 ms and returned 4 rows
Query 7 iteration 2 took 9324.8 ms and returned 4 rows
Query 7 avg time: 10017.75 ms
```

```
Query 19 iteration 0 took 2099.1 ms and returned 1 rows
Query 19 iteration 1 took 1907.6 ms and returned 1 rows
Query 19 iteration 2 took 1910.4 ms and returned 1 rows
Query 19 avg time: 1972.36 ms
```