@Dandandan 2022-10-15
Hm getting different results now - may have been something running in the background on my machine - will update the results accordingly